Feature/selectable pages (#5)
* [FEATURE] Use ordinary pages as mail template

This adds a new page type "Email" and a addtional
ce for form results. This allows the user
to compose a template for later use in the
the form email finisher.

* [TASK] Add pages to dropdown dynamically

All pages of type 125 (Email) are displayed
in the Finisher select field.

This enables the freedom to choose or
create any template and fill it with content
according to a customers needs.

* [TASK] Prefill general variables, remove custom template

* [TASK] Optimize plaintext rendering

Adds a page type to generate plaintext
for email.

* [TASK] Validate given identifier

The given identifier may only contain
characters, numbers, - and _. It is also
supposed to be unique.

A validation has been added to ensure a
given string is a valid and unique identifier.

* [TASK] Do not update stage when identifier ist not valid

A invalid identifier can break the gui e.g. quotes.

This is fixed by only updating the stage area
when a validator is valid and not already
in use.

* [TASK] Add identifier field to form elements

https://github.com/b13/form-custom-templates/issues/7

* [TASK] Optimize Formatting of plaintext mail part

https://github.com/b13/form-custom-templates/issues/8

* [TASK] Use page type 101 for plaintext.

https://github.com/b13/form-custom-templates/issues/8

* [TASK] Remove ctypes for doktype "Email" (125)

https://github.com/b13/form-custom-templates/issues/8

* [TASK] Add normalized params to fill existing markers

This also gives access to additional information in the template.

https://github.com/b13/form-custom-templates/issues/8

* [TASK] Add ts to configure path to the resultList template

* [TASK] Disable save buttons while stage is not yet updated

https://github.com/b13/form-custom-templates/issues/10

* [TASK] Add v10 compatability again

* [TASK] Add details about template configuration

* [TASK] Add preview button to email template pages

* [TASK] Make typeNum and doktype configurable via extension settings

* [TASK] Make typeNum and doktype configurable via extension settings

* Feature/testing (#12)

* [TASK] Add basic testing scaffold
* [TASK] Update phpstan to lates, fix return types
* [TASK] Add phpstan baseline, activate acceptance tests
* [TASK] Make phpstan happy in all php versions
* [TASK] Remove PreviewBuilder for v10 compatability, do not run tests on v10 for now
* [TASK] Add acceptance tests